requirements.txt
Flask==3.0.0
APScheduler==3.10.4
supabase==2.23.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1